PowerQuant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "powerquant"
version = "0.1.0"
description = "Power consumption prediction using quantile regression"
authors = [{name = "Anonymous"}]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "optuna>=3.0.0",
    "catboost>=1.2.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
powerquant = "build:cli"

[project.optional-dependencies]
fvcore = ["fvcore>=0.1.5.post20221221"]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]

[tool.uv.sources]

[tool.hatch.build.targets.wheel]
packages = ["build.py", "www"]
